+++ src/Laba/Habitat/IBehaviour.java
@@ -3,6 +3,6 @@ package Laba.Habitat;
 public interface IBehaviour {
     void setX(int x);
     void setY(int y);
-    double getX();
-    double getY();
+    int getX();
+    int getY();
 }
